+++ src/index.ts
@@ -2,6 +2,15 @@
 export {
   useFluent,
   GrammyFluentOptions,
+
+} from './middleware';
+
+export {
   FluentContextFlavor,
 
-} from './fluent-grammy';
+} from './context';
+
+export {
+  LocaleNegotiator,
+
+} from './locale-negotiator';