+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 
-# 0.0.1
+# 0.0.2 (minor release)
+
+- introduced customized locale negotiation via `LocaleNegotiator`
+- added `localeNegotiator` property
+- added `defaultLocale` property
+- `en` locale is now used as a fallback value
+- updated and extended the documentation
+
+
+# 0.0.1 (initial release)
 
 Initial release.